autoposter/requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
schedule==1.2.0
SQLAlchemy==2.0.25
python-crontab==2.6.1
